refactor: split router.rs and task_runner.rs into handler modules
Split oversized files to prepare for VibeGuard absorption:

- router.rs (851→112 lines): now pure match dispatch, delegates to
  handler modules for all RPC method implementations
- task_runner.rs (528→250 lines): retains types, TaskStore, and
  spawn_task; execution logic moved to task_executor.rs
- task_executor.rs (187 lines): run_task and review loop logic
- handlers/{thread,gc,skills,observe,rules,exec}.rs: domain-grouped
  handler functions extracted from the router match arms

All 93 existing tests pass with no behavior change.
