+++ crates/harness-server/src/handlers/mod.rs
@@ -1,0 +1,6 @@
+pub mod exec;
+pub mod gc;
+pub mod observe;
+pub mod rules;
+pub mod skills;
+pub mod thread;
